Backend/requirements.txt
fastapi
uvicorn
torch
transformers
pillow
numpy
requests
python-multipart 